Update and refactor new class.
